Application profile specific view added
