+++ datamodel-ui/public/locales/sv/common.json
@@ -136,6 +136,7 @@
   "no-upper-attributes": "",
   "no-upper-classes": "",
   "not-added": "",
+  "not-defined": "",
   "own-information": "",
   "page-indicator-text": "",
   "page-number": "",
